@@ -3,19 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@react-spring/web": "^9.5.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "poweraudio": "^0.1.8",
-    "react": "^17.0.0",
+    "powerglitch": "^2.0.0",
+    "react": "^18.2.0",
     "react-device-detect": "^2.2.2",
-    "react-dom": "^17.0.0",
-    "react-scripts": "5.0.0",
-    "react-scroll-parallax": "^3.0.6",
-    "react-spring": "^9.4.4",
-    "react-waypoint": "^10.1.0",
-    "sass": "^1.50.0",
-    "web-vitals": "^2.1.4"
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "react-scroll-parallax": "^3.3.1",
+    "react-spring": "^9.5.2",
+    "react-waypoint": "^10.3.0",
+    "sass": "^1.54.9",
+    "web-vitals": "^3.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -42,8 +44,8 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.4",
-    "postcss": "^8.4.12",
-    "tailwindcss": "^3.0.24"
+    "autoprefixer": "^10.4.8",
+    "postcss": "^8.4.16",
+    "tailwindcss": "^3.1.8"
   }
 }
